Replace Ubuntu torrent file in test suite
Ubuntu invalidates some of its provided downloads as the versions
expire and new ones are released.  The removed torrent was one such
download that was removed from service.  In exploration, a successful
connection was met with a cryptic error message.

Switching the test torrent file to something much smaller, as the Ubuntu
torrent pieces were also cumbersome to manipulate in the REPL.  New
torrent sample file is hosted by archive.org and should be available for
a greater length of time.